+++ spec/dicey/distribution_calculators/auto_selector_spec.rb
@@ -35,8 +35,8 @@ module Dicey
     context "with a large list of small regular dice" do
       let(:dice) { RegularDie.from_count(800, 3) }
 
-      it "returns RegularMultinomialCoefficients" do
-        expect(selected_calculator).to be_a DistributionCalculators::RegularMultinomialCoefficients
+      it "returns RegularMultinomial" do
+        expect(selected_calculator).to be_a DistributionCalculators::RegularMultinomial
       end
     end
 
@@ -83,11 +83,11 @@ module Dicey
     end
 
     context "if initialized with a custom list of calculators" do
-      let(:selector) { described_class.new([DistributionCalculators::RegularMultinomialCoefficients.new]) }
+      let(:selector) { described_class.new([DistributionCalculators::RegularMultinomial.new]) }
       let(:dice) { RegularDie.from_count(2, 2) }
 
       it "considers only the given calculators" do
-        expect(selected_calculator).to be_a DistributionCalculators::RegularMultinomialCoefficients
+        expect(selected_calculator).to be_a DistributionCalculators::RegularMultinomial
       end
 
       context "if no calculators are compatible" do